feat(基础): 用户、角色、部门、菜单、字典完成

+++ src/core/sys/user/userRole.service.ts
@@ -1,7 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { CreateQueryParams } from '@nestjsx/crud-request'
 import { SysUserRole } from './userRole.entity';
+import R, { RType } from 'src/utils/R';
+import { SysUserRoleUpdate } from './user.controller';
+
 
 @Injectable()
 export class SysUserRoleService {
@@ -12,25 +16,21 @@ export class SysUserRoleService {
     //
   }
 
-  async getUserRoleList(id: string): Promise<SysUserRole> {
-    let value = { id: 0 };
-    const result = await this.repo
-      .createQueryBuilder()
-      .where('user_id = :id', { id })
-      .getOne();
-    if (result) {
-      value = result;
-    }
-    return value;
+  async getUserRoleList(id: string, query: CreateQueryParams): Promise<RType<SysUserRole[]>> {
+    return R.list<SysUserRole>(this.repo, {
+      page: query.page,
+      limit: query.limit,
+      params: `user_id=${id}`
+    })
   }
 
-  async setUserRole(body: SysUserRole): Promise<SysUserRole> {
+  async setUserRole(body: SysUserRoleUpdate): Promise<boolean> {
     // 存在先删除
-    await this.repo
-      .createQueryBuilder()
-      .where('user_id = :id', { id: body.userId })
-      .delete()
-      .execute();
-    return this.repo.save(body);
+    await this.repo.createQueryBuilder().where(`user_id = ${body.userId}`).delete().execute()
+    for (let i = 0; i < body.data.length; i += 1) {
+      const item = body.data[i];
+      await this.repo.save(item);
+    }
+    return true;
   }
 }
